Add files via upload
Overall, I was able to achieve all of the required items on the grading rubrik.  There may be a cleaner, more efficient way to do this depending on the 3 or 6 hole prompt, but it works and feel pretty good about it overall.  Was able to put this Deliverable together pretty quickly.